requirements.txt
Django==1.11.12
django-contrib-comments==1.8.0
djangorestframework==3.8.2
factory-boy==2.10.0
Faker==0.8.13
python-dateutil==2.7.2
pytz==2018.4
six==1.11.0
text-unidecode==1.2
